+++ MyWebApplication/pom.xml
@@ -9,13 +9,13 @@
 		<version>2.6.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.food</groupId>
+	<groupId>com.example</groupId>
 	<artifactId>MyWebApplication</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>MyWebApplication</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>11</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
