+++ MyWebApplication/src/main/java/com/application/food/repository/ProductRepository.java
@@ -1,0 +1,12 @@
+package com.application.food.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.application.food.models.Product;
+
+public interface ProductRepository extends CrudRepository<Product, Integer> {
+	
+	
+
+}
+
