--- MyWebApplication/src/main/java/com/food/application/MyWebApplication.java
@@ -1,15 +1,0 @@
-package com.food.application;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan({ "com.food.MyWebApplication", "com.food.controller" })
-public class MyWebApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MyWebApplication.class, args);
-	}
-
-}
