--- MyWebApplication/src/main/java/com/food/repository/ProductRepository.java
@@ -1,13 +1,0 @@
-package com.food.repository;
-
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.food.entities.Product;
-
-public interface ProductRepository extends CrudRepository<Product, String> {
-	
-	
-
-}
-
