+++ MyWebApplication/src/main/resources/application.properties
@@ -1,4 +1,8 @@
-server.port=9000
-spring.jpa.show-sql=true
+spring.datasource.url = jdbc:mysql://localhost:3306/foodexplorer
+spring.datasource.username = root
+spring.datasource.password = Adithy@9989
+spring.jpa.show-sql = true
 spring.h2.console.enabled=true
-
+spring.jpa.hibernate.ddl-auto = create
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+server.port=9090
